Now with Poetry™! - Added Poetry files for ease of development - Moved syntax highlighting code into a separate file with its own command - Converted to using argparse - Made rich a dependency and always enabled - Moved various files around to make the code into an actual package
